R4E UI JUnits improvement2 (part 2)
We still have stability problems with the automated UI tests where some
of them fail occasionally.

This patch:

- De-'singleton'izes TestUtils and R4UITestMain so we avoid potential
simultaneous changes in the test workspace and in the shared repository

- Introduces R4ETestCase to minimize the code impacts of the
de-'singleton'ization on the actual tests (which have to extend
R4ETestCase)

- Ensures that the workspace projects have a unique name per test suite

- Ensures that the shared drive is also unique per test suite

- Moved the share drive(s) from /tmp to the workspace so it is
guaranteed to be cleaned up event if a test case bombed with an
exception.

- Remove the now obsolete R4ETestSetup class


Change-Id: I12a24e70924d65becfcaa012b41143eeaf98d977
Signed-off-by: Francois Chouinard <fchouinard@gmail.com>